requirements.txt
Flask==2.3.3
mysqlclient==2.1.1
qrcode==7.4.2
Pillow==10.0.0
flask-cors==4.0.0